separated slider as it's own component
