+++ src/web/src/components/graphComponent/chartComponent/chartComponent.jsx
@@ -53,10 +53,6 @@ const options = {
 		legend: {
 			position: "top",
 		},
-		title: {
-			display: true,
-			text: "Chart.js Line Chart",
-		},
 	},
 };
 const data = {
